style: fix stan issues

+++ app/Services/Transactions/Aggregates/Fees/Historical/DayAggregate.php
@@ -18,7 +18,7 @@ final class DayAggregate
     {
         return $this->mergeWithPlaceholders(
             (new RangeAggregate())->aggregate(Carbon::now()->subDay(), Carbon::now(), 'H'),
-            $this->placeholders(Carbon::now()->subDay()->timestamp + 3600, Carbon::now()->timestamp + 3600, 3600, 'H')->take(24)
+            $this->placeholders((int) Carbon::now()->subDay()->timestamp + 3600, (int) Carbon::now()->timestamp + 3600, 3600, 'H')->take(24)
         );
     }
 }
